+++ samples/Sentry.Samples.ME.Logging/Program.cs
@@ -19,11 +19,11 @@ using var loggerFactory = LoggerFactory.Create(builder =>
         // Optionally configure options: The default values are:
         options.MinimumBreadcrumbLevel = LogLevel.Information; // It requires at least this level to store breadcrumb
         options.MinimumEventLevel = LogLevel.Error; // This level or above will result in event sent to Sentry
-        options.MinimumLogLevel = LogLevel.Trace; // This level or above will result in log sent to Sentry
+        options.ExperimentalLogging.MinimumLogLevel = LogLevel.Trace; // This level or above will result in log sent to Sentry
 
-        // This option enables the (experimental) Sentry Logs.
-        options.EnableLogs = true;
-        options.SetBeforeSendLog(static log =>
+        // This option enables Logs sent to Sentry.
+        options.Experimental.EnableLogs = true;
+        options.Experimental.SetBeforeSendLog(static log =>
         {
             log.SetAttribute("attribute-key", "attribute-value");
             return log;