+++ src/Sentry/Extensibility/DisabledHub.cs
@@ -33,7 +33,19 @@ public class DisabledHub : IHub, IDisposable
     /// <summary>
     /// No-Op.
     /// </summary>
+    public void ConfigureScope<TArg>(Action<Scope, TArg> configureScope, TArg arg)
+    {
+    }
+
+    /// <summary>
+    /// No-Op.
+    /// </summary>
     public Task ConfigureScopeAsync(Func<Scope, Task> configureScope) => Task.CompletedTask;
+
+    /// <summary>
+    /// No-Op.
+    /// </summary>
+    public Task ConfigureScopeAsync<TArg>(Func<Scope, TArg, Task> configureScope, TArg arg) => Task.CompletedTask;
 
     /// <summary>
     /// No-Op.