+++ src/Sentry/TransactionTracer.cs
@@ -386,11 +386,11 @@ public class TransactionTracer : IBaseTracer, ITransactionTracer
 
         // Clear the transaction from the scope and regenerate the Propagation Context
         // We do this so new events don't have a trace context that is "older" than the transaction that just finished
-        _hub.ConfigureScope(scope =>
+        _hub.ConfigureScope(static (scope, transactionTracer) =>
         {
-            scope.ResetTransaction(this);
+            scope.ResetTransaction(transactionTracer);
             scope.SetPropagationContext(new SentryPropagationContext());
-        });
+        }, this);
 
         // Client decides whether to discard this transaction based on sampling
         _hub.CaptureTransaction(new SentryTransaction(this));